requirements.txt
numpy==2.0.2
pandas==2.2.2
python-igraph==0.11.9
pyvista==0.45.2
scipy